requirements.txt
gunicorn
wheel
lightgbm
numpy
pandas
flask
mecab-python3
umap-learn
tqdm